set up redux structure
